# crates/matric-api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Fortémi API
  description: |
    AI-enhanced knowledge base with semantic search, automatic linking, and NLP pipelines.

    ## Features
    - **Notes**: CRUD operations for markdown notes with AI revision
    - **Search**: Hybrid search (FTS + semantic + RRF fusion) with multilingual support
    - **SKOS Taxonomy**: W3C SKOS concept schemes, concepts, and collections
    - **Collections**: Hierarchical folder organization for notes
    - **Templates**: Note templates with variable substitution
    - **Versioning**: Dual-track version history (original + revised)
    - **Embeddings**: Configurable embedding sets with MRL support
    - **Document Types**: 131 pre-configured types with auto-detection
    - **Archives**: Named archive containers with retention policies
    - **Graph**: Recursive semantic link exploration
    - **Backup**: Knowledge shards, database backups, and knowledge archives
    - **Observability**: Knowledge health dashboard, timeline, activity feeds
    - **Jobs**: Background NLP pipeline monitoring
    - **OAuth2**: Full OAuth2 with Dynamic Client Registration (RFC 7591)

    ## Authentication
    The API supports two authentication methods:
    - **OAuth2**: Authorization Code, Client Credentials, and Refresh Token grants
    - **API Keys**: Simple Bearer token authentication for trusted integrations

    ## Search Query Syntax
    ```
    hello world        # Match all words (AND)
...
    apple -orange      # Exclude word
...

...
servers:
...
security:
  - bearerAuth: []
  - oauth2: []
...
paths:
...
  /api/v1/document-types/{name}:
    get:
      summary: Get document type
      operationId: getDocumentType
      tags: [Document Types]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentType'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update document type
      operationId: updateDocumentType
      tags: [Document Types]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                filename_patterns:
                  type: array
                  items:
                    type: string
                chunk_strategy:
                  type: string
                  enum: [semantic, syntactic, fixed, none]
                chunk_size:
                  type: integer
                chunk_overlap:
                  type: integer
      responses:
        '204':
          description: Updated

    delete:
      summary: Delete custom document type
      operationId: deleteDocumentType
      tags: [Document Types]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '403':
          description: Cannot delete built-in type
...
  /api/v1/archives/{name}:
    get:
      summary: Get archive
      operationId: getArchive
      tags: [Archives]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archive details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Archive'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update archive
      operationId: updateArchive
      tags: [Archives]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                retention_policy:
                  type: string
      responses:
        '204':
          description: Updated

    delete:
      summary: Delete archive
      operationId: deleteArchive
      tags: [Archives]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    # Notes
...
    DocumentType:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        filename_patterns:
          type: array
          items:
            type: string
        magic_content_patterns:
          type: array
          items:
            type: string
        chunk_strategy:
          type: string
          enum: [semantic, syntactic, fixed, none]
        chunk_size:
          type: integer
        chunk_overlap:
          type: integer
        is_builtin:
          type: boolean
...
    Archive:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        retention_policy:
          type: string
          nullable: true
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token (OAuth2 access token or API key)

    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read: Read access to notes, tags, and search
            write: Create and update notes and tags
            delete: Delete notes
            admin: Administrative access
            mcp: MCP server access (includes read and write)
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to notes, tags, and search
            write: Create and update notes and tags
            delete: Delete notes
            admin: Administrative access
            mcp: MCP server access (includes read and write)

    clientCredentials:
      type: http
      scheme: basic
      description: Client credentials (client_id:client_secret)